Change frequency to seasonal frequency

Change frequency to seasonal frequency

Usage

htr_seasonal_frequency(indir, tempdir, outdir, months, months_name)

Arguments

indir

Directory where input files are located

tempdir

Temporary directory where specific months are selected

outdir

Directory where output files will be saved

months

Define season (based in the numbered format of months)

months_name

Define seasone name for the filename's suffix

Examples

if (FALSE) { # \dontrun{
htr_seasonal_frequency(indir = here("data", "proc", "sliced", "omip", variable),
tempdir = here("data", "temporary"), # you can also have this as a folder that isn't temporary, you just need to delete the line of code below that deletes all the files in this folder. `tempdir` holds the files that selects the particular months of interest for the specific season
outdir = here("data", "proc", "seasonal", "omip", variable),
months = c("01", "02", "03"), # define season (in numbered format)
months_name = "jan-mar" # define season name
)
} # }