Slice Period

Slice Period

Usage

htr_slice_period(
  indir,
  outdir,
  freq,
  scenario,
  year_start,
  year_end,
  overwrite = TRUE
)

Arguments

indir

Directory where input files are located

outdir

Directory where output files will be saved

freq

The temporal frequency to be used in the analysis

scenario

The CMIP scenario to be used in the analysis.

year_start

Starting year

year_end

Ending year

overwrite

Should the output files be overwritten if they already exist (defaults to TRUE)

Author

Dave Schoeman and Tin Buenafe

Examples

if (FALSE) { # \dontrun{
htr_slice_period(
  indir = file.path(base_dir, "data", "proc", "merged", "tos"), # input directory
outdir = file.path(base_dir, "data", "proc", "sliced", "tos"), # output directory
freq = "Omon", # ocean, daily
scenario = "ssp",
year_start = 2020,
year_end = 2100,
overwrite = FALSE
)
} # }